Minimizes the weights of the neural network using backward propagation.
This requires flattening the arrays and reshaping the output. Originally used the scipy.optimize.fmin_bfgs function but this did not converge in a reasonable time, even by limiting maxiter. Switched to scipy.optimize.fmin_l_bfgs_b a local, bounded bfgs implementation. This converged much more quickly when adjusting the maxiter and maxfun (the maximum number of iterations and function calls in the minimization).