Normalizes the features by subtracting the mean and dividing by the standard deviation.
This is the same as the file in linear regression, except that I switched the order of the indices. I created my own index order in the linear regression folder, and the standard for the coursera machine learning course is to have the features in the second dimension. Since this makes the matrix multiplication standardized, I will continue to use features as the second dimension (and training examples in the first dimension).