Made the plot of the error curve much more functional and readable.
Added axis labels and legend. More importantly, limited the maximum y-value plotted to be twice the largest median of the two types of error (training set and cross validation errors). This makes plots much more readable as the largest error tends to dominate the plot range but is not a very interesting feature in the plot.